# agents/mia/output/openapi.yaml
openapi: "3.1.0"
info:
  title: Tokenfly Agent Team Lab — Dashboard API
  version: "1.0.0"
  description: |
    REST API for the Tokenfly Agent Team Lab dashboard.
    Provides access to agent status, task management, messaging, announcements,
    team channel, search, and CEO inbox.

    **Base URL:** `http://localhost:3100`

    **Rate Limits:**
    - Read endpoints: 120 requests/minute per IP
    - Write endpoints: 20 requests/minute per IP

  contact:
    name: Mia (API Engineer)
servers:
  - url: http://localhost:3100
    description: Local dev server
  - url: http://localhost:3199
    description: E2E test server
...
paths:
...
  /api/agents/{name}:
    get:
      tags: [Agents]
      summary: Get agent detail
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/AgentName"
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List all tasks
      operationId: listTasks
      parameters:
        - name: assignee
          in: query
          required: false
          description: Filter by assignee name (case-insensitive)
          schema:
            type: string
          example: bob
        - name: status
          in: query
          required: false
          description: Filter by task status (case-insensitive)
          schema:
            type: string
            enum: [open, in_progress, blocked, in_review, done, cancelled]
          example: open
      responses:
        "200":
          description: Array of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
              example:
                - id: "1"
                  title: "Architecture Review"
                  description: "Review system architecture"
                  priority: high
                  assignee: alice
                  status: in_progress
                  created: "2026-03-29"
                  updated: "2026-03-29"

    post:
      tags: [Tasks]
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
            example:
              title: "Implement rate limiting"
              description: "Add per-IP rate limiting to all write endpoints"
              priority: high
              assignee: mia
      responses:
        "201":
          description: Task created — returns full task object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
              example:
                id: 12
                title: "Implement rate limiting"
                description: "Add per-IP rate limiting to all write endpoints"
                priority: high
                assignee: mia
                status: open
                created: "2026-03-29"
                updated: "2026-03-29"
        "400":
          description: Bad request — missing title, whitespace-only title, or invalid priority
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingTitle:
                  value:
                    error: "title is required"
                invalidPriority:
                  value:
                    error: "priority must be one of: low, medium, high, critical"
...
  /api/tasks/{id}:
    patch:
      tags: [Tasks]
      summary: Update a task
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
            example:
              status: done
      responses:
        "200":
          description: Task updated — returns full updated task object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
              example:
                id: 3
                title: "QA Review"
                description: "Run test suites"
                priority: high
                assignee: tina
                status: done
                created: "2026-03-29"
                updated: "2026-03-29"
        "400":
          description: Invalid status or priority value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidStatus:
                  value:
                    error: "status must be one of: open, in_progress, blocked, in_review, done, cancelled"
                invalidPriority:
                  value:
                    error: "priority must be one of: low, medium, high, critical"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Tasks]
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 3
      responses:
        "200":
          description: Task deleted — returns the removed task
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    AgentName:
      name: name
      in: path
      required: true
      description: Agent name (alphanumeric, hyphens, underscores only)
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
      examples:
        alice:
          value: alice
        bob:
          value: bob
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "missing title"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "agent not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AgentDetail:
      type: object
      required: [name, alive, inbox]
      description: |
        Agent detail returned by GET /api/agents/:name (backend/api.js).
        The inbox field contains metadata only — no message content is exposed
        to prevent unauthenticated data exposure (QI-003).
      properties:
        name:
          type: string
        alive:
          type: boolean
          description: True if heartbeat was updated within the last 5 minutes
        heartbeat_at:
          type: string
          format: date-time
          nullable: true
          description: ISO timestamp of last heartbeat file modification
        statusMd:
          type: string
          nullable: true
          description: Raw content of status.md (also present as status_md)
        status_md:
          type: string
          nullable: true
          description: Alias for statusMd
        inbox:
          type: array
          description: Inbox file metadata — filenames and read status only (no content)
          items:
            $ref: "#/components/schemas/InboxMessageMeta"
    InboxMessageMeta:
      type: object
      description: Inbox message metadata (QI-003 — content not exposed)
      properties:
        file:
          type: string
          description: Filename in the chat_inbox directory
          example: "2026_03_29_22_00_00_from_ceo.md"
        read:
          type: boolean
          description: True if filename starts with "read_"
...
    Task:
      type: object
      properties:
        id:
          type: string
          description: Task ID (numeric string from markdown table)
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignee:
          type: string
        status:
          type: string
          enum: [open, in_progress, blocked, in_review, done, cancelled]
        created:
          type: string
          format: date
        updated:
          type: string
          format: date
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        assignee:
          type: string
          maxLength: 50
    UpdateTaskRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignee:
          type: string
        status:
          type: string
          enum: [open, in_progress, blocked, in_review, done, cancelled]

    # --- Announcements ---
...